# Conflicts:
#	src/main/java/com/swp391/e_Motion_be/config/WebSecurityConfig.java

+++ src/main/java/com/swp391/e_Motion_be/service/user/UserService.java
@@ -1,27 +1,65 @@
 package com.swp391.e_Motion_be.service.user;
 
-import com.swp391.e_Motion_be.dto.requests.user.ChangePasswordUserRequest;
-import com.swp391.e_Motion_be.dto.requests.user.UpdateProfileRequest;
-import com.swp391.e_Motion_be.dto.responses.UserResponse;
-import com.swp391.e_Motion_be.entity.User;
+import com.swp391.e_Motion_be.dto.requests.user.*;
+import com.swp391.e_Motion_be.dto.responses.stats.*;
+import com.swp391.e_Motion_be.dto.responses.user.PageAndFilterUserResponse;
+import com.swp391.e_Motion_be.dto.responses.user.StaffStatsResponse;
+import com.swp391.e_Motion_be.dto.responses.user.UserResponse;
+import com.swp391.e_Motion_be.entity.*;
+import com.swp391.e_Motion_be.enums.CheckType;
 import com.swp391.e_Motion_be.enums.ErrorCode;
+import com.swp391.e_Motion_be.enums.RentalStatus;
+import com.swp391.e_Motion_be.enums.Role;
+import com.swp391.e_Motion_be.enums.vehicle.VehicleStatus;
 import com.swp391.e_Motion_be.exception.AppException;
+import com.swp391.e_Motion_be.mapper.RentalMapper;
+import com.swp391.e_Motion_be.mapper.ReservationMapper;
 import com.swp391.e_Motion_be.mapper.UserMapper;
-import com.swp391.e_Motion_be.repository.UserRepository;
+import com.swp391.e_Motion_be.repository.*;
+import com.swp391.e_Motion_be.service.StaffService;
+import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDate;
+import java.time.LocalDateTime;
+import java.time.Month;
 import java.util.List;
+import java.util.Map;
+import java.util.stream.Collectors;
+import java.util.stream.IntStream;
 
 @Service
 @RequiredArgsConstructor
 public class UserService {
     private final UserRepository userRepository;
+    private final VehicleRepository vehicleRepository;
+    private final ReservationRepository reservationRepository;
+    private final RentalRepository rentalRepository;
+    private final RentalCheckListRepository rentalCheckListRepository;
+
     private final UserMapper userMapper;
     private final BCryptPasswordEncoder passwordEncoder;
+    private final StationRepository stationRepository;
+    private final ReservationMapper reservationMapper;
+    private final RentalMapper rentalMapper;
+    private final StaffRepository staffRepository;
+    private final StaffService staffService;
+
+    public User currentUser(){
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String email = ((User) authentication.getPrincipal()).getEmail();
+
+        return userRepository.findByEmail(email)
+                .orElseThrow(() -> new AppException(ErrorCode.USER_NOT_EXISTS));
+    }
 
     public List<UserResponse> getAllUsers() {
         return userRepository.findAll().stream()
@@ -40,13 +78,20 @@ public class UserService {
     }
 
     public UserResponse getUserByEmail(String email) {
-        return userMapper.toUserResponse(userRepository.findByEmail(email)
-                .orElseThrow(() -> new AppException(ErrorCode.USER_NOT_EXISTS)));
+        User user = userRepository.findByEmail(email)
+                .orElseThrow(() -> new AppException(ErrorCode.USER_NOT_EXISTS));
+
+        return userMapper.toUserResponse(user);
     }
 
     public void deleteUserByEmail(String email) {
         User user = userRepository.findByEmail(email)
                 .orElseThrow(() -> new AppException(ErrorCode.USER_NOT_EXISTS));
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String currentUserEmail = ((User) authentication.getPrincipal()).getEmail();
+        if(user.getEmail().equals(currentUserEmail)) {
+            throw new AppException(ErrorCode.CANNOT_DELETE_OWN_ACCOUNT);
+        }
         userRepository.delete(user);
     }
 
@@ -62,9 +107,8 @@ public class UserService {
             throw new AppException(ErrorCode.OLD_PASSWORD_NOT_MATCH);
         }
 
-        // Check confirm password
-        if (!input.getNewPassword().equals(input.getConfirmNewPassword())) {
-            throw new AppException(ErrorCode.CONFIRM_PASSWORD_NOT_MATCH);
+        if(input.getOldPassword().equals(input.getNewPassword())) {
+            throw new AppException(ErrorCode.NEW_PASSWORD_SAME_AS_OLD);
         }
 
         // Save encoded new password
@@ -73,12 +117,22 @@ public class UserService {
     }
 
 
+    public List<UserResponse> getUserByEmailContains(String email) {
+        return userRepository.findByEmailContains(email).stream()
+                .map(userMapper::toUserResponse)
+                .toList();
+    }
+
     public UserResponse updateProfile(UpdateProfileRequest input) {
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
         String email = ((User) authentication.getPrincipal()).getEmail();
 
         User user = userRepository.findByEmail(email)
                 .orElseThrow(() -> new AppException(ErrorCode.USER_NOT_EXISTS));
+        // Check phone mới có tồn tại chưa
+        if(!user.getPhone().equals(input.getPhone()) && userRepository.existsByPhone(input.getPhone())) {
+            throw new AppException(ErrorCode.PHONE_EXITS);
+        }
 
         user.setFullName(input.getFullName());
         user.setPhone(input.getPhone());
@@ -86,5 +140,217 @@ public class UserService {
         userRepository.save(user);
 
         return userMapper.toUserResponse(user);
+    }
+
+    public PageAndFilterUserResponse findByPageAndFilterAndSearch(PageAndFilterUserRequest request) {
+        User user = currentUser();
+        List<Role> roleList = (request.getRoleList() == null || request.getRoleList().isEmpty())
+                ? List.of(Role.ROLE_USER, Role.ROLE_ADMIN, Role.ROLE_STAFF)
+                : request.getRoleList();
+
+        List<Boolean> blockedList = (request.getBlockedList() == null || request.getBlockedList().isEmpty())
+                ? List.of(true, false)
+                : request.getBlockedList();
+
+        Pageable pageable = PageRequest.of(request.getPage() - 1, request.getLimit(), Sort.by("id").ascending());
+        Page<User> userPage;
+
+        if(user.getRole() == Role.ROLE_STAFF){
+            Station station = stationRepository.findById(user.getStaff().getStation().getId())
+                    .orElseThrow(() -> new AppException(ErrorCode.STATION_NOT_FOUND));
+            userPage = userRepository.findUsersByStaff(blockedList, roleList, request.getSearch(), station.getId(), pageable);
+        }else if(request.getStationId() != null) {
+            Station station = stationRepository.findById(request.getStationId())
+                    .orElseThrow(() -> new AppException(ErrorCode.STATION_NOT_FOUND));
+            userPage = userRepository.findByBlockedInAndRoleInAndEmailContainsAndStaff_Station_Id(blockedList, List.of(Role.ROLE_STAFF), request.getSearch(), station.getId(), pageable);
+        }else{
+            userPage = userRepository.findByBlockedInAndRoleInAndEmailContains(blockedList, roleList, request.getSearch(), pageable);
+        }
+
+        List<UserResponse> users = userPage.getContent().stream()
+                .map(userMapper::toUserResponse)
+                .toList();
+
+        return new PageAndFilterUserResponse(users, userPage.getTotalPages());
+    }
+
+    public TotalStatsResponse getTotalStatsDashboard(){
+        List<Rental> rentals = rentalRepository.findAll();
+        List<Vehicle> vehicles = vehicleRepository.findAll();
+
+        long totalUsers = userRepository.countByRole(Role.ROLE_USER);
+        long totalVehicles = vehicles.size();
+        long totalReservations = reservationRepository.count();
+        double totalRevenue = calculateRevenue(rentals);
+        double usageRate = calculateUsageRate(vehicles);
+
+        return new TotalStatsResponse(
+                totalUsers,
+                totalVehicles,
+                totalReservations,
+                rentals.size(),
+                totalRevenue,
+                usageRate);
+    }
+
+    public List<StationStatsResponse> getStationDetailDashboard(){
+        List<Station> stations = stationRepository.findAll();
+        return stations.stream().map(station -> {
+            List<Rental> rentals = rentalRepository.findByStation_Id(station.getId());
+
+            long totalVehicles = vehicleRepository.countByStation_Id(station.getId());
+            double usageRate = calculateUsageRate(vehicleRepository.findByStation_Id(station.getId()));
+            double revenue = calculateRevenue(rentals);
+            List<Integer> peakHours = getPeakHours(rentals);
+
+            return new StationStatsResponse(
+                    station.getName(),
+                    revenue,
+                    totalVehicles,
+                    rentals.size(),
+                    usageRate,
+                    peakHours
+            );
+        }).toList();
+    }
+
+    public List<RevenueResponse> getRevenueInYearDashboard(){
+        List<Rental> rentals = rentalRepository.findAll().stream()
+                .filter(r -> r.getStatus() == RentalStatus.COMPLETED)
+                .filter(r -> r.getEndTime().getYear() == LocalDateTime.now().getYear())
+                .toList();
+
+        Map<Integer, Double> revenueByMonth = rentals.stream()
+                .collect(Collectors.groupingBy(r -> r.getEndTime().getMonth().getValue(),
+                        Collectors.summingDouble(Rental::getRentFee)));
+
+        return IntStream.rangeClosed(1, 12)
+                .mapToObj(month -> new RevenueResponse(
+                        "T" + Month.of(month).getValue(), // tên tháng (JANUARY, FEBRUARY,...)
+                        revenueByMonth.getOrDefault(month, 0.0)
+                ))
+                .toList();
+    }
+
+    public List<PeakHourResponse> getPeakHoursInMonthDashboard(){
+        Map<Integer, Long> hourFrequency = rentalRepository.findAll().stream()
+                .filter(rental -> (rental.getStatus() == RentalStatus.COMPLETED || rental.getStatus() == RentalStatus.ONGOING))
+                .filter(r -> {
+            LocalDate date = r.getStartTime().toLocalDate();
+            return date.getMonthValue() == LocalDate.now().getMonthValue() && date.getYear() == LocalDate.now().getYear();
+        })
+                .collect(Collectors.groupingBy(rental -> rental.getStartTime().getHour(), Collectors.counting()));
+
+        return IntStream.rangeClosed(1, 23)
+                .mapToObj(hour -> new PeakHourResponse(
+                        hour,
+                        hourFrequency.getOrDefault(hour, 0L)
+                ))
+                .toList();
+    }
+
+    public List<Integer> getPeakHours(List<Rental> rentals){
+        Map<Integer, Long> hourFrequency = rentals.stream()
+                .filter(rental -> (rental.getStatus() == RentalStatus.COMPLETED || rental.getStatus() == RentalStatus.ONGOING))
+                .filter(r -> r.getEndTime().getYear() == LocalDateTime.now().getYear())
+                .collect(Collectors.groupingBy(rental -> rental.getStartTime().getHour(), Collectors.counting()));
+        long maxCount = hourFrequency.values().stream()
+                .max(Long::compareTo).orElse(0L);
+        return hourFrequency.entrySet().stream()
+                .filter(entry -> entry.getValue() == maxCount)
+                .map(Map.Entry::getKey)
+                .toList();
+    }
+
+    // Helper để tính tỷ lệ sử dụng
+    private double calculateUsageRate(List<Vehicle> vehicles) {
+        if (vehicles.isEmpty()) return 0;
+        long inUse = vehicles.stream().filter(v -> v.getStatus() == VehicleStatus.ONGOING || v.getStatus() == VehicleStatus.CHECKING).count();
+        return ((double) inUse / vehicles.size()) * 100;
+    }
+
+    // Helper để tính tổng doanh thu
+    private double calculateRevenue(List<Rental> rentals) {
+        return rentals.stream()
+                .filter(r -> r.getStatus() == RentalStatus.COMPLETED)
+                .filter(r -> r.getEndTime().getYear() == LocalDateTime.now().getYear())
+                .mapToDouble(Rental::getRentFee)
+                .sum();
+    }
+
+    public DataAdminDashboard getDataAdminDashboard(){
+        TotalStatsResponse totalStats = getTotalStatsDashboard();
+        List<StationStatsResponse> stationDetails = getStationDetailDashboard();
+        List<RevenueResponse> revenueInYear = getRevenueInYearDashboard();
+        List<PeakHourResponse> peakHours = getPeakHoursInMonthDashboard();
+
+        return new DataAdminDashboard(totalStats, stationDetails, revenueInYear, peakHours);
+    }
+
+    public UserResponse createUserByAdmin(@Valid CreateUserRequest request) {
+        if(userRepository.existsByEmail(request.getEmail())) {
+            throw new AppException(ErrorCode.EMAIL_ALREADY_EXISTS);
+        }
+        if(userRepository.existsByPhone(request.getPhone())) {
+            throw new AppException(ErrorCode.PHONE_EXITS);
+        }
+
+        User newUser = new User();
+        newUser.setFullName(request.getFullName());
+        newUser.setEmail(request.getEmail());
+        newUser.setPhone(request.getPhone());
+        newUser.setPassword(passwordEncoder.encode(request.getPassword()));
+        newUser.setRole(request.getRole());
+        if(request.getRole() == Role.ROLE_STAFF) {
+            if(request.getStationId() != null){
+                Staff staff = staffService.createStaff(newUser, request.getStationId());
+                newUser.setStaff(staff);
+            }else{
+                throw new AppException(ErrorCode.STATION_NOT_FOUND);
+            }
+        }
+        newUser.setEnabled(true);
+        newUser.setBlocked(false);
+
+        userRepository.save(newUser);
+
+        return userMapper.toUserResponse(newUser);
+    }
+
+    public void toggleStatusUser(String email) {
+        User user = userRepository.findByEmail(email)
+                .orElseThrow(() -> new AppException(ErrorCode.USER_NOT_EXISTS));
+        user.setBlocked(!user.isBlocked());
+        userRepository.save(user);
+    }
+
+    public void updateUserByAdmin(UpdateUserRequest request) {
+        User user = userRepository.findByEmail(request.getEmail())
+                .orElseThrow(() -> new AppException(ErrorCode.USER_NOT_EXISTS));
+        user.setRole(request.getRole());
+        if(request.getRole() == Role.ROLE_STAFF) {
+            if (request.getStationId() != null) {
+                staffService.updateStaff(user.getEmail(), request.getStationId());
+            } else {
+                throw new AppException(ErrorCode.STATION_NOT_FOUND);
+            }
+        } else {
+            staffRepository.findByUser_EmailAndIsDeleteFalse(request.getEmail())
+                    .ifPresent(staff -> {
+                        staff.setDelete(true);
+                        staffRepository.save(staff);
+                    });
+        }
+        userRepository.save(user);
+    }
+
+    public StaffStatsResponse getTransactions(Long staffId) {
+        Staff staff = staffRepository.findById(staffId)
+                .orElseThrow(() -> new AppException(ErrorCode.STAFF_NOT_FOUND));
+
+        int deliveries = rentalCheckListRepository.countByStaff_IdAndType(staff.getId(), CheckType.CHECK_IN);
+        int pickups = rentalCheckListRepository.countByStaff_IdAndType(staff.getId(), CheckType.CHECK_OUT);
+
+        return new StaffStatsResponse(deliveries, pickups);
     }
 }
